+++ tests/test08/run.sh
@@ -185,8 +185,8 @@ if [ "$CONDUCTOR" = ".TRUE." ] ; then
    $WANT_BIN/conductor.x < $TEST_HOME/conductor.in > $TEST_HOME/conductor.out
    if [ ! -e CRASH ] ; then 
       echo "$ECHO_T done" 
-      mv dos.dat $TEST_HOME/dos.dat
-      mv cond.dat $TEST_HOME/cond.dat
+      test -e dos.dat && mv dos.dat $TEST_HOME/dos.dat
+      test -e cond.dat && mv cond.dat $TEST_HOME/cond.dat
    else
       echo "$ECHO_T problems found" ; cat CRASH ; exit 1
    fi
